refactor(physics): decouple monolithic nbody simulation
Implement a strict Separation of Concerns (SoC) for the N-Body physics
engine. The legacy monolithic nbody.c/.h has been decommissioned and
split into specialized internal modules, preserving the exact same
public API for external consumers (e.g., scene_nbody.c).

- API & Types: nbody.h (facade) and nbody_types.h (data structures).
- Physics: nbody_physics.c (Symplectic Verlet integration & Softening).
- Diagnostics: nbody_diagnostics.c (Hamiltonian energy drift validation).
- Presets: nbody_preset.c (Orbiter initialization and configurations).
- Render Bridge: nbody_render.c (Instanced rendering mapping).
- Build: Update CMakeLists.txt to compile the new module and rigorously
  enforce IEEE 754 compliance (-fno-fast-math) across the new files.
